+++ utils/ai.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"regexp"
 	"strings"
+
 	"wechat-robot-client/vars"
 )